add onclick function in organization card
