+++ app/(private-layout)/get-started/_components/OrganizationCard/index.module.css
@@ -13,4 +13,5 @@
     color: #a57ced;
     border-color: #a57ced;
   }
+  cursor: pointer;
 }
