+++ app/(private-layout)/get-started/_components/OrganizationCard/index.tsx
@@ -1,14 +1,26 @@
+"use client";
+
 import { Avatar, Box, Flex, Text } from "@mantine/core";
 import React from "react";
-import { OrganizationNavBtn } from "../../_viewModules/OrganizationNavBtn";
+// import { OrganizationNavBtn } from "../../_viewModules/OrganizationNavBtn";
 import classes from "./index.module.css";
 import Image from "next/image";
 import { OrganizationType as OrganizationCardPropsType } from "../../types";
+import { GET_STARTED_PAGE } from "@/constants/dataTestId";
+import { DSM_APP_URL } from "@/constants/commons";
+import { handleOrganizationNav } from "../../actions";
+import { DsmIconButton } from "@/components/DsmIconButton";
 
 export const OrganizationCard = (props: OrganizationCardPropsType) => {
-  const { id, name, logoUrl } = props;
+  const { id: orgId, name, logoUrl } = props;
+
+  const handleClick = async () => {
+    await handleOrganizationNav(orgId);
+    window.location.href = DSM_APP_URL;
+  };
+
   return (
-    <Box className={classes.card}>
+    <Box className={classes.card} onClick={handleClick}>
       <Flex gap={"sm"} justify={"space-between"} align={"center"}>
         <Flex direction={"column"} gap={4}>
           {logoUrl ? (
@@ -33,7 +45,15 @@ export const OrganizationCard = (props: OrganizationCardPropsType) => {
             {name}
           </Text>
         </Flex>
-        <OrganizationNavBtn orgId={id} className={classes.icon} />
+        <DsmIconButton
+          data-testid={`${GET_STARTED_PAGE.NAV_ORGANIZATION_BTN}-${orgId}`}
+          size={"32px"}
+          color="white"
+          iconProps={{ icon: "arrowRight", size: 18 }}
+          radius={"50%"}
+          variant="outline"
+          className={classes.icon}
+        />
       </Flex>
     </Box>
   );
